mcp/chemical-kg-mvp/requirements.txt
# Core dependencies - Fixed for compatibility issues
langchain==0.1.0
langchain-community==0.0.20
# Fix for SQLite version compatibility with ChromaDB
pysqlite3-binary
# Use consistent pydantic version
pydantic==1.10.13
# Fix tokenizers build issue
tokenizers==0.19.1
sentence-transformers>=2.3.0
PyMuPDF==1.23.8
pymupdf-fonts==1.0.5
Pillow>=9.0.0,<11.0.0
rdkit==2023.9.3
streamlit==1.29.0
pandas==2.0.3
numpy>=1.2.0

# API and networking
requests==2.31.0
aiohttp==3.9.1
urllib3==2.1.0

# LLM related
ollama==0.1.7
openai==1.6.1
tiktoken==0.5.2

# Vector store and embeddings
faiss-cpu==1.7.4
hnswlib==0.8.0

# Chemistry specific
pubchempy==1.0.4
chembl-webresource-client==0.10.8
molvs==0.1.1
# DECIMER packages (TensorFlow provided by base image)
# Note: TensorFlow 2.12.0-gpu already included in base image
DECIMER==2.7.1
decimer-segmentation==1.4.0
# Additional DECIMER dependencies
pystow
pillow-heif
efficientnet
selfies
pyyaml
scikit-image>=0.2.0
pdf2image
scipy
matplotlib

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6
opencv-python-headless==4.8.1.78
psutil==5.9.6

# Development tools (compatible with older Python)
ipython==8.12.3
black==22.12.0
pytest==7.2.0
pytest-mock==3.10.0